Restructure code files Add ConfigEnumSetup for value lists Refactor methods in ConfigSetup to use setup
